Move ColorRangeSelector and ColorRange into visualizations
Every consumer of both components lives in visualizations, so they move
out of metabase/common: ColorRangeSelector and ColorRange become
siblings under visualizations/components/settings, imported relatively.
The registration line in register.ts now pins ColorRangeSelector to the
ChartSettingColorRangeProps contract with a satisfies check, so the
component and the contract can no longer drift apart silently.
