+++ frontend/src/metabase/visualizations/components/settings/ColorRange/ColorRange.stories.tsx
