+++ frontend/src/metabase/visualizations/components/settings/ColorRange/index.ts
