+++ frontend/src/metabase/visualizations/components/settings/ColorRangeSelector/ColorRangePopover.tsx
