+++ frontend/src/metabase/visualizations/components/settings/ColorRangeSelector/ColorRangeSelector.stories.tsx
