+++ frontend/src/metabase/visualizations/components/settings/ColorRangeSelector/ColorRangeSelector.unit.spec.tsx
