+++ frontend/src/metabase/visualizations/components/settings/ColorRangeSelector/ColorRangeToggle.tsx
