+++ frontend/src/metabase/visualizations/components/settings/ColorRangeSelector/ColorRangeToggle.unit.spec.tsx
