+++ frontend/src/metabase/visualizations/components/settings/ColorRangeSelector/index.ts
